% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{adam_block}
\alias{adam_block}
\title{Adam block}
\usage{
adam_block(...)
}
\arguments{
\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_data_block()}}}
}
\description{
Block listing pharmaversesdtm
}
