% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{new_adam_block}
\alias{new_adam_block}
\title{Adam block}
\usage{
new_adam_block(selected = character(), ...)
}
\arguments{
\item{selected}{Default selected dataset. Must belong to the
pharmaverseadam package.}

\item{...}{Passed to \code{\link[blockr:new_block]{blockr::new_block()}}}
}
\description{
Block listing pharmaversesdtm
}
